Finished cookies port by adding openID support. --------- Enable OpenID auth in the embedded webkit browser.
Extract openid_url and openid_token tokens from the login response in process_login_success_response() and send them to
LLViewerMedia if they're present.

Added LLViewerMedia::openIDSetup() to receive openid_url and openid_token, and added code to LLViewerMedia to do a POST with
LLHTTPClient, retrieve the resulting cookie, and push it into the central cookie store. Also made sure the OpenID cookie gets
re-added when the cookie store is cleared.

Added LLPluginCookieStore::setCookiesFromHost() to properly add a cookie that may not have a domain set. Made
LLPluginCookieStore::Cookie::parse() add missing domain and path fields to cookies as necessary.

Fixed an issue where carriage returns in the string passed to LLPluginCookieStore::setCookies() or
LLPluginCookieStore::setCookiesFromHost() would cause a parse failure.
---------
Fix for incorrect handling of an OpenID url containing a port number

The code was using the "authority" part of the URL returned by LLURL when constructing the cookie's domain, which incorrectly
included the port number. This change makes sure any port number and username/password that might be in the authority are
stripped from the string before handing it to the cookie code.